+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.2.1] - Unreleased
 
+### Added
+
+- Add alias for verify method as scan.
+- Add configuration for auto_contact and verify_scan_statuses.
+
+### Changed
+
+- Ensure that events are only created if the status has changed or if the status is in the verify_scan_statuses list and verification fails.
+
 ## [0.2.0] - 2025-02-26
 
 ### Changed